Fix file:// URI parsing
A previous change to fix #1344 to handle # characters in literal no-scheme
paths broke file:// scheme URI parsing by including the entire literal URI in
the file path, which breaks encoded paths (like those with spaces) and URIs
with non-path data (such as query parameters).

This partially reverts the prior fix, so that # is not supported in file://
URIs, requiring the user to either use a no-scheme URI or encode the path,
as # is reserved in URIs for fragment declarations.

Fixes: #1513

Test: FileUriMapperTest#ignoresAfterPathCorrectly
Test: FileUriMapperTest#decodesEncodedPath