Fix transformation with default response in FakeImageLoaderEngine (#2884)
* Fix transformation with default response in FakeImageLoaderEngine

When using the default interceptor for a default response, the chain used was not the transformed one.
This caused the Request to not have a NONE Transition.Factory, and causing tests to still running transitions.

For example this caused issues with our Paparazzi tests where, requests with default result where still having the crossfade animation causing the tests to be flaky.

* run spotlessApply